[clean] separates socket tests from unit tests
